@@ -24,8 +24,8 @@
         "rollerworks/version": "^0.2.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^5.6",
-        "symfony/phpunit-bridge": "^3.4",
+        "phpunit/phpunit": "^7.3.1",
+        "symfony/phpunit-bridge": "^4.1",
         "symfony/debug": "^3.4"
     },
     "autoload": {
